Auto creation of roles

+++ src/main/java/org/ieeervce/gatekeeper/entity/User.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -36,7 +36,7 @@ public class User {
 
 
     @Column(nullable = false, unique = true)
-    private  long number;
+    private  BigInteger number;
     @CreationTimestamp
     @Column(nullable = false)
     private LocalDateTime createdAt;
@@ -81,7 +81,7 @@ public class User {
     public void setUserId(Long userId) {
         this.userId = userId;
     }
-    public void setNumber(long number) {
+    public void setNumber(BigInteger number) {
         this.number = number;
     }
 
@@ -133,7 +133,7 @@ public class User {
         this.updatedAt = updatedAt;
     }
 
-    public long getNumber() {
+    public BigInteger getNumber() {
         return number;
     }
